+++ src/wildboar/explain/counterfactual/_helper.py
@@ -32,7 +32,7 @@ def _proximity(
     metric : str or callable, optional
         The distance metric
 
-        See ``_DISTANCE_MEASURE.keys()`` for a list of supported metrics.
+        See ``_METRICS.keys()`` for a list of supported metrics.
 
     metric_params: dict, optional
         Parameters to the metric.